Refactor: update file structure
