+++ 06 - Type Ahead/index-START.html
@@ -15,6 +15,6 @@
         <li>or a state</li>
       </ul>
     </form>
-    <script src="./script.js"></script>
+    <script src="./script.js" type="module"></script>
   </body>
 </html>
