+++ 06 - Type Ahead/utils/findCityMatches.js
@@ -1,0 +1,6 @@
+export const findCityMatches = (searchWord = '', cities) =>
+  cities.filter(
+    (city) =>
+      city.city.toLowerCase().includes(searchWord) ||
+      city.state.toLowerCase().includes(searchWord)
+  );
